Refactor AppHeader and Sidebar components to use NewNoteButton
- Replaced the inline new note button implementation in both AppHeader and Sidebar with a reusable NewNoteButton component.
- Removed unnecessary imports and code related to note creation logic from Sidebar, streamlining the component's functionality.
